+++ CHANGELOG.md
@@ -65,6 +65,7 @@
   - html: allow `class`, `title` and `style` attributes
   - dropdown: close menu when receiving a value
   - dropdown: removed empty 1st option
+  - switch: removed `showValues` (inconsistent with menu/dropdown, feature covered by `values` property)
 
 - widget removals
   - `push`, `toggle`: merged into `button`